Bug fixes and new unit tests.
Bug fixes
---------
If a general Exception was caught (i.e.: the API can't connect to the database) in any of the endpoints, a 200 HTTP code would be returned.
What should happen is that the endpoints should return a 500 HTTP code, indicating a internal server error.

Replaced all the abort() calles with statemens that return JSON with a status code.
The function, create_JSON_response(), was created for this purpose.

New unit tests
--------------
Added tests for HTTP 500 responses from the endpoints.
Added a test for HTTP 404 response for the /file/{filename} endpoint.